Fix an integer overflow when starting a search 25 days after the previous one

+++ dcpp/SearchManager.h
@@ -85,7 +85,7 @@ public:
 	void onRES(const AdcCommand& cmd, const UserPtr& from, const string& removeIp = Util::emptyString);
 
 	int32_t timeToSearch() {
-		return 5 - (static_cast<int32_t>(GET_TICK() - lastSearch) / 1000);
+		return 5 - (static_cast<int64_t>(GET_TICK() - lastSearch) / 1000);
 	}
 
 	bool okToSearch() {
